Application of Eigenvalues and Eigenvectors
Multiple components describing some structure can be organized as a single vector. If the state of the structure is changing in time due to some process, the vector is referred as a state vector. Such dynamic processes are often modeled at discrete times calculating the state vector as a linear transformation of a previous state vector. This model is called a discrete dynamical system.

Starting with an initial state  𝑋0
 , the transition to the next state  𝑋1
  is a linear transformation defined with a matrix  𝑃
 :  𝑋1=𝑃𝑋0
 . That leads to  𝑋2=𝑃𝑋1=𝑃2𝑋0
 ,  𝑋3=𝑃3𝑋0
 , and so on. This implies that  𝑋𝑡=𝑃𝑋𝑡−1
  for  𝑡=0,1,2,3,…
 .

A discrete dynamical system can be used to model browsing web pages. Web pages often contain links to other pages, for simplicity assume that the browser is only following links to a new page rather than navigating to an unlinked one.

Let the probabilities that the browser is on a particular page at time  𝑡
  be a state vector  𝑋𝑡
 . Navigation from one page to another advances the model from one state vector  𝑋𝑡−1
  to another state vector  𝑋𝑡
 . A linear transformation, defined by a matrix  𝑃
 , will have entries  𝑎𝑖𝑗
  with the probabilities that the browser navigates to page  𝑖
  from page  𝑗
 . For fixed column  𝑗
 , the entries represent a probability distribution describing location of the browser at the next step. Thus, the entries in each column must add to one.